+++ pom.xml
@@ -55,7 +55,7 @@
         <dependency>
           <groupId>org.apache.avro</groupId>
           <artifactId>avro</artifactId>
-          <version>1.8.1</version>
+          <version>1.8.2</version>
         </dependency>
         <dependency>
           <groupId>com.twitter</groupId>
@@ -98,8 +98,8 @@
          <artifactId>maven-compiler-plugin</artifactId>
          <version>2.3.2</version>
          <configuration>
-             <source>1.7</source>
-             <target>1.7</target>
+             <source>1.8</source>
+             <target>1.8</target>
          </configuration>
      </plugin>
   </plugins>
